backend/pyproject.toml
[project]
name = "stock-analysis-backend"
version = "1.0.0"
description = "Stock Analysis API with Tushare Pro"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "akshare>=1.12.62",
    "pydantic>=2.5.3",
    "pandas>=2.1.4",
    "tushare>=1.4.29",
    "python-dotenv>=1.0.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "deepagents>=0.5.0a4",
    "tavily>=1.1.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "Pillow>=10.0.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
trusted-host = "pypi.tuna.tsinghua.edu.cn"
dev-dependencies = []
